AIGC_Project/requirements.txt
torch
 torchvision
 pillow
 opencv-python
 numpy